api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# OCR
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Vector Database
marqo==3.14.0
sentence-transformers==2.2.2

# LLM and LangChain - Compatible versions
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0

# Data Processing - Updated for Python 3.12 compatibility
numpy>=1.26.0
pandas>=2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1